BUILD/DEBUG: lru: fix printf format in debug code
cppcheck reports in issue #1184 a type mismatch between "%d" and the
unsigned int "misses" in the standalone debug code of lru.c. Let's
switch to "%u".

+++ src/lru.c
@@ -295,7 +295,7 @@ int main(int argc, char **argv)
 		ret += get_value(lru, statistical_prng() & 65535);
 	}
 	/* just for accuracy control */
-	printf("ret=%llx, hits=%d, misses=%d (%d %% hits)\n", ret, total-misses, misses, (int)((float)(total-misses) * 100.0 / total));
+	printf("ret=%llx, hits=%u, misses=%u (%d %% hits)\n", ret, (unsigned)(total-misses), misses, (int)((float)(total-misses) * 100.0 / total));
 
 	while (lru64_destroy(lru));
